Algebraic degree of semidefinite programming

A list of the values of the algebraic degree of semidefinite programming delta(m,n,r) for n=2,3,...,14. The values are given as strings to avoid rounding errors.

alg_deg_data

Format

A list of 13 character matrices; the rows and columns corresponds to m and r, respectively.

Source

The degrees have been computed with Maple using a formula given in https://arxiv.org/abs/1509.05241, https://projecteuclid.org/euclid.kmj/1478073765.

See also

alg_deg

Package: symconivol