Converting samples of index constrained eigenvalues to samples of bivariate chi-bar-squared distribution

constr_eigval_to_bcbsq converts samples of the index constrained eigenvalue distribution to samples of the corresponding bivariate chi-bar-squared distribution.

constr_eigval_to_bcbsq(samp)

Arguments

samp
a list containing (a selection of) the elements \'ep\', \'ef\', \'en\', which are matrices (of the same length) whose rows give the sampled eigenvalues

Value

The output of constr_eigval_to_bcbsq is a two-column matrices whose rows correspond to the samples from the bivariate chi-bar-squared distribution.

See also

constr_eigval, prepare_em_cm, estim_em_cm

Package: symconivol

Examples

# NOT RUN {
library(tidyverse)
library(rstan)

filename <- "tmp.stan"
M <- constr_eigval( beta=2, n=12, ind_low=8, ind_upp=9, filename=filename )
stan_samp <- stan( file = filename, data = M$data,
                   chains = 1, warmup = 1e3, iter = 1e5, cores = 2, refresh = 1e4 )
file.remove(filename)

m_samp <- constr_eigval_to_bcbsq( rstan::extract(stan_samp) )
# }