Exact curvature measures of symmetric cones

curv_meas_exact returns the exact values of the curvature measures of symmetric cones, which are known (n==1,2,3).

curv_meas_exact(beta, n)

Arguments

beta
Dyson index specifying the underlying (skew-) field:

beta==1:

real numbers

beta==2:

complex numbers

beta==4:

quaternion numbers

n
size of matrix.

Value

The output of curv_meas_exact is a list of six elements:

A: the combined matrix of curvature measures; A is given in terms of the other parameters by (A_const + A_sqrt2*sqrt(2) + A_piinv/pi + A_sqrt2_pi*sqrt(2)/pi)/denom

A_const: integer matrix for the constant term

A_sqrt2: integer matrix for the sqrt(2) term

A_piinv: integer matrix for the 1/pi term

A_sqrt2_pi: integer matrix for the sqrt(2)/pi term

denom: common denominator of all terms

Details

The known curvature measures are elements in the ring Q[sqrt(2),1/pi], so the exact values can be given in terms of integers corresponding to a common denominator and corresponding integer matrices for the coefficients of the natural expansion in 1, sqrt(2), 1/pi, sqrt(2)/pi. These matrices are returned by this function, along with the denominator and the combined matrix of curvature measures.

See also

alg_deg

Package: symconivol

Examples

# considering the case of 3x3 complex unitary matrices
CM <- curv_meas_exact(2,3)

# sum of intrinsic volumes is equal to one
sum( CM$A )

#> [1] 1

# sum of even (and odd) index intrinsic volumes is 1/2
sum( CM$A %*% rep_len(c(1,0),dim(CM$A)[2]) )

#> [1] 0.5

# A is given by combining the remaining matrices and the denominator
norm( CM$A - ( CM$A_const + CM$A_sqrt2*sqrt(2) + CM$A_piinv/pi + CM$A_sqrt2_pi*sqrt(2)/pi )/CM$denom )

#> [1] 0