Pataki bounds

pat_bnd provides the Pataki inequalities for given beta and n.

pat_bnd(beta, n)

Arguments

beta
Dyson index specifying the underlying (skew-) field:

beta==1:

real numbers

beta==2:

complex numbers

beta==4:

quaternion numbers

n
size of matrix

Value

The output of pat_bnd is a list of five elements:

d: ambient dimension of corresponding Euclidean space, given in terms of beta,n by n+beta*choose(n,2)

k_low: function returning the lower bound for index k, given r

k_upp: function returning the upper bound for index k, given r

r_low: function returning the lower bound for index r, given k

r_upp: function returning the upper bound for index r, given k

See also

leigh

Package: symconivol

Examples

# considering the case of 3x3 complex unitary matrices
beta <- 2
n <- 3
P <- pat_bnd(beta,n)

# ambient dimension
P$d

#> [1] 9

# bounds for index k, given r
P$k_low(0:n)

#> [1] 0 1 4 9

P$k_upp(0:n)

#> [1] 0 5 8 9

# bounds for index r, given k
P$r_low(0:P$d)

#>  [1] 0 1 1 1 2 2 2 2 2 3

P$r_upp(0:P$d)

#>  [1] 0 1 1 1 1 1 2 2 2 3