Evaluate bivariate chi-bar-squared samples for maximum likelihood estimation

prepare_em_cm takes a two-column matrix whose rows form iid samples from a bivariate chi-bar-squared distribution and prepares the data used in maximum likelihood estimation.

prepare_em_cm(d, low, upp, m_samp)

Arguments

d
the dimension of the bivariate chi-bar squared distribution.

low
lower bound for k; has to be >0

upp
upper bound for k; has to be <d

m_samp
two-column matrix whose rows from iid samples from a bivariate chi-bar-squared distribution.

Value

The output of prepare_em_cm is (low-upp+1) row matrix whose kth row contains the products of the density values of the chi_k^2 and chi_(d-k)^2 distributions evaluated in the sample points; the row-form of the matrix is more convenient for the computations.

Details

This function works pretty much exactly as prepare_em from the conivol package, the only difference being that the "boundary cases" k==0,n do not have to be considered/are ignored. In the general case this is not needed, but for the curvature measures this is a useful feature.

See also

prepare_em, constr_eigval, constr_eigval_to_bcbsq, estim_em_cm

Package: symconivol

Examples

CM <- curv_meas_exact(4,3)$A[,2]
CM <- CM/sum(CM)

m_samp <- conivol::rbichibarsq(1e5,CM)

str( prepare_em_cm( 15, 1, 9, m_samp ))

#>  num [1:9, 1:100000] 0.031951 0.017709 0.005994 0.001523 0.000312 ...